Fix PATH variable overriding.
This fucking took me 45 minutes of testing, but I finally found it out.
Line 47 of this fucking .zshrc was overriding all over exports of PATH
at the top of the file.

I literally have no idea how I am so stupid or how that was added, but
fuck this entire earth, this was so annoying to debug.